requirements.txt
langchain
langchain-community>=0.0.10
pyyaml>=6.0.2
perplexity-client
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.5.0
langchain-perplexity
langchain-openai

